+++ tests/integration/src/test/java/org/apache/pulsar/tests/integration/io/sources/debezium/PulsarDebeziumSourcesTest.java
@@ -29,7 +29,6 @@ import org.apache.pulsar.common.policies.data.TenantInfoImpl;
 import org.apache.pulsar.common.schema.SchemaInfo;
 import org.apache.pulsar.tests.integration.containers.DebeziumMongoDbContainer;
 import org.apache.pulsar.tests.integration.containers.DebeziumMySQLContainer;
-import org.apache.pulsar.tests.integration.containers.DebeziumOracleDbContainer;
 import org.apache.pulsar.tests.integration.containers.DebeziumPostgreSqlContainer;
 import org.apache.pulsar.tests.integration.io.PulsarIOTestBase;
 import org.testcontainers.shaded.com.google.common.collect.Sets;
@@ -62,11 +61,6 @@ public class PulsarDebeziumSourcesTest extends PulsarIOTestBase {
     @Test(groups = "source")
     public void testDebeziumMongoDbSource() throws Exception{
         testDebeziumMongoDbConnect("org.apache.kafka.connect.json.JsonConverter", true);
-    }
-
-    @Test(groups = "source", timeOut = 1200000)
-    public void testDebeziumOracleDbSource() throws Exception{
-        testDebeziumOracleDbConnect("org.apache.kafka.connect.json.JsonConverter", true);
     }
 
     private void testDebeziumMySqlConnect(String converterClassName, boolean jsonWithEnvelope) throws Exception {
@@ -195,43 +189,6 @@ public class PulsarDebeziumSourcesTest extends PulsarIOTestBase {
         runner.testSource(sourceTester);
     }
 
-    private void testDebeziumOracleDbConnect(String converterClassName, boolean jsonWithEnvelope) throws Exception {
-        final String tenant = TopicName.PUBLIC_TENANT;
-        final String namespace = TopicName.DEFAULT_NAMESPACE;
-        final String outputTopicName = "debe-output-topic-name-" + testId.getAndIncrement();
-        final String consumeTopicName = "debezium/oracle/XE.INV.CUSTOMERS";
-        final String sourceName = "test-source-debezium-oracle-" + functionRuntimeType + "-" + randomName(8);
-
-        // This is the event count to be created by prepareSource.
-        final int numMessages = 1;
-
-        @Cleanup
-        PulsarClient client = PulsarClient.builder()
-                .serviceUrl(pulsarCluster.getPlainTextServiceUrl())
-                .build();
-
-        @Cleanup
-        PulsarAdmin admin = PulsarAdmin.builder().serviceHttpUrl(pulsarCluster.getHttpServiceUrl()).build();
-        initNamespace(admin);
-
-        admin.topics().createNonPartitionedTopic(consumeTopicName);
-        admin.topics().createNonPartitionedTopic(outputTopicName);
-
-        @Cleanup
-        DebeziumOracleDbSourceTester sourceTester = new DebeziumOracleDbSourceTester(pulsarCluster);
-        sourceTester.getSourceConfig().put("json-with-envelope", jsonWithEnvelope);
-
-        // setup debezium oracle server
-        DebeziumOracleDbContainer debeziumOracleDbContainer = new DebeziumOracleDbContainer(pulsarCluster.getClusterName());
-        sourceTester.setServiceContainer(debeziumOracleDbContainer);
-
-        PulsarIODebeziumSourceRunner runner = new PulsarIODebeziumSourceRunner(pulsarCluster, functionRuntimeType.toString(),
-                converterClassName, tenant, namespace, sourceName, outputTopicName, numMessages, jsonWithEnvelope,
-                consumeTopicName, client);
-
-        runner.testSource(sourceTester);
-    }
-
     protected void initNamespace(PulsarAdmin admin) {
         log.info("[initNamespace] start.");
         try {
@@ -242,7 +199,6 @@ public class PulsarDebeziumSourcesTest extends PulsarIOTestBase {
                 "debezium/mysql-avro",
                 "debezium/mongodb",
                 "debezium/postgresql",
-                "debezium/oracle"
             };
             Policies policies = new Policies();
             policies.retention_policies = new RetentionPolicies(-1, 50);
