+++ tests/integration/src/test/resources/pulsar-io-ora-source.xml
@@ -19,12 +19,10 @@
 
 -->
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<!-- TODO: we have to put suite files in one file to avoid executing TESTNG test suites multiple times.
-           see {@link https://github.com/cbeust/testng/issues/508} -->
-<suite name="Pulsar Test Suite" parallel="instances" thread-count="1">
-    <suite-files>
-        <suite-file path="./pulsar-io-sinks.xml" />
-        <suite-file path="./pulsar-io-sources.xml" />
-        <cuite-file path="./pulsar-io-ora-source.xml" />
-    </suite-files>
+<suite name="Pulsar IO Oracle Source Integration Tests" verbose="2" annotations="JDK">
+    <test name="pulsar-function-process-test-suite" preserve-order="true" >
+        <classes>
+            <class name="org.apache.pulsar.tests.integration.io.sources.debezium.PulsarDebeziumOracleSourceTest" />
+        </classes>
+    </test>
 </suite>
