+++ tests/integration/src/test/resources/pulsar.xml
@@ -35,5 +35,6 @@
         <suite-file path="./pulsar-backwards-compatibility.xml" />
         <suite-file path="./pulsar-io-sinks.xml" />
         <suite-file path="./pulsar-io-sources.xml" />
+        <suite-file path="./pulsar-io-ora-source.xml" />
     </suite-files>
 </suite>
